Organized some rules from rules.py
- CV01 renamed
- LT03 renamed
- Organized rule scripts and test cases
- updated __init__.py
