+++ src/sqlfluff_easy_ql/__init__.py
@@ -19,15 +19,17 @@ def get_rules() -> List[Type[BaseRule]]:
     is available for the validation steps in the meta class.
     """
     # i.e. we DO recommend importing here:
-    from sqlfluff_easy_ql.rules import (
-        Rule_EasyQL_L002,
-        Rule_EasyQL_L003
-    )  # noqa: F811
-    from sqlfluff_easy_ql.LT02 import Rule_EasyQL_LT02
     from sqlfluff_easy_ql.LT01 import Rule_EasyQL_LT01
+    from sqlfluff_easy_ql.LT02 import Rule_EasyQL_LT02
+    from sqlfluff_easy_ql.LT03 import Rule_EasyQL_LT03
+    from sqlfluff_easy_ql.CV01 import Rule_EasyQL_CV01
 
-    return [Rule_EasyQL_LT01, Rule_EasyQL_LT02,
-            Rule_EasyQL_L002, Rule_EasyQL_L003]
+    return [
+        Rule_EasyQL_LT01,
+        Rule_EasyQL_LT02,
+        Rule_EasyQL_LT03,
+        Rule_EasyQL_CV01
+    ]
 
 
 @hookimpl
